requirements.txt
logger_setup==0.17
python-dotenv==1.1.0
Requests==2.32.3
ffmpeg-python
scenedetect[opencv]
accelerate
opencv-python
pillow
nvidia-ml-py3
rich
sentence-transformers
torch
torchvision
torchaudio
transformers
accelerate
bitsandbytes
qwen-vl-utils
qwen-vl-utils[decord]
numpy
pymediainfo